+++ app/Http/Controllers/SupporterController.php
@@ -15,15 +15,18 @@ use App\Models\LR;
 
 use App\Notifications\User2SupporterRegisteredNotification;
 
+use App\Http\Requests\SupporterRequest;
+use App\Http\Requests\User2SupporterRequest;
+
 class SupporterController extends Controller
 {
-    //サポーターページ
+    //サポーターマイページ
     public function index(Request $request){
     
         //サポーターかどうかのチェック
         Gate::authorize('supporter');
 
-       //m_hogoshaが紐づけがない場合には、informationを表示
+       //m_supporterが紐づけがない場合には、informationを表示
        $args=[];
        if(Gate::allows('supporter-nobind')){
             $args=['alertInfo'=>MessageConst::NOT_BINDED,];
@@ -32,45 +35,40 @@ class SupporterController extends Controller
     }
 
     /*システム管理者が使用する画面*/
-    //サポーター登録画面へ
-    public function add(Request $request){
-        $mode = 'add';
-        $item = new Supporter;
 
-        //クエリ文字列にSupporterCdがついている場合は、編集モードで開く
-        if(isset($request->SupporterCd)){
-            $item = Supporter::where('SupporterCd',$request->SupporterCd)->first();
-            $mode = 'edit';
-        }
-
-        $items = Supporter::all();
-        $lrs = LR::all();
-
-        //リダイレクト時には、セッションにalertが入ってくる可能性があるので拾う
-        $alertComp='';
-        if($request->session()->has('alertComp')){
-            $alertComp = $request->session()->get('alertComp');
-        }
-        $alertErr='';
-        if($request->session()->has('alertErr')){
-            $alertErr = $request->session()->get('alertErr');
-        }
+    //GET マスタ一覧画面へ
+    public function list(Request $request){
         
-        $arg=[
-            'mode'=>$mode,
-            'item'=>$item,
-            'items'=>$items,
-            'lrs' =>$lrs,
-            'alertComp'=>$alertComp,
-            'alertErr'=>$alertErr,
+        $args=[
         ];
 
-        return view('Supporter.add',$arg);
+        return view('Supporter.list',$args)->with(parent::getAlertSessions());
 
     }
-    //サポーター登録画面のPOST
-    public function create(Request $request){
-        $this->validate($request, Supporter::$rules_create);
+    //GET マスタ編集画面へ
+    public function edit(Request $request){
+        $mode = 'create';
+
+        //supporterCdがあれば、編集モード、なければ新規モード
+        if($request->has('supporterCd')){
+            $mode='update';
+        }
+        
+        $args=[
+            'mode'=>$mode,
+            'createAction' =>"/supporter/create",
+            'updateAction' =>"/supporter/update",
+            'deleteAction' =>"/supporter/delete",
+            'backURL'=>"/supporter/list/",
+        ];
+
+        return view('Supporter.edit',$args)->with(parent::getAlertSessions());
+
+    }
+    
+    //POST Create処理
+    public function create(SupporterRequest $request){
+        // $this->validate($request, Supporter::$rules_create);
         $supporter = new Supporter;
         $form = $request->all();
         unset($form['_token']);
@@ -93,12 +91,13 @@ class SupporterController extends Controller
         $args=[
         ];
 
-        return redirect()->route('supporter-add',$args)->with('alertComp',MessageConst::ADD_COMPLETED);
+        return redirect()->route('supporter-list',$args)->with('alertComp',MessageConst::ADD_COMPLETED);
 
     }
-    public function edit(Request $request){
-        $this->validate($request, Supporter::$rules_edit);
-        $supporter = Supporter::where('SupporterCd',$request->SupporterCd)->first();
+    // POST Update処理
+    public function update(SupporterRequest $request){
+        // $this->validate($request, Supporter::$rules_edit);
+        $supporter = Supporter::find($request->SupporterCd);
         $form = $request->all();
         unset($form['_token']);
         unset($form['create']);
@@ -120,46 +119,60 @@ class SupporterController extends Controller
         $args=[
         ];
 
-        return redirect()->route('supporter-add',$args)->with('alertComp',MessageConst::EDIT_COMPLETED);
+        return redirect()->route('supporter-list',$args)->with('alertComp',MessageConst::EDIT_COMPLETED);
 
     }
+    // POST Delete処理
     public function delete(Request $request){
 
+        $supporter = Supporter::find($request->SupporterCd);
+
+        $supporter->delete();
+
+        //登録後の再取得
+        $args=[
+        ];
+
+        return redirect()->route('supporter-list',$args)->with('alertComp',MessageConst::DELETE_COMPLETED);
     }
-    //user2サポーター登録画面へ /user2suppo/add/
-    public function u2sadd(Request $request, Response $response){
+
+    /* ここからUser2Hogosha */
     
-        $items = User2Supporter::getu2sData();
+    //user2サポーター登録第一画面へ /user2suppo/list/
+    public function u2slist(Request $request){
 
-        $itemsSupporter = Supporter::all();
+        $args=[
+        ];
 
-        //リダイレクト時には、セッションにalertが入ってくる可能性があるので拾う
-        $alertComp='';
-        if($request->session()->has('alertComp')){
-            $alertComp = $request->session()->get('alertComp');
-        }
-        $alertErr='';
-        if($request->session()->has('alertErr')){
-            $alertErr = $request->session()->get('alertErr');
+        return view('user2suppo.list',$args)->with(parent::getAlertSessions());
+
+    }
+    //user2サポーター登録第二画面へ /user2suppo/edit/ GET
+    public function u2sedit(Request $request){
+        $mode = 'create';
+
+        //u2h_idがあれば、編集モード、なければ新規モード
+        if($request->has('u2s_id')){
+            //当マスタは編集モード非対応のため
+            $mode='delete';
         }
         
         $args=[
-            'items'=>$items,
-            'itemsSupporter' =>$itemsSupporter,
-            'alertComp'=>$alertComp,
-            'alertErr'=>$alertErr,
+            'mode'=>$mode,
+            'createAction' =>"/user2suppo/create",
+            'updateAction' =>"",
+            'deleteAction' =>"/user2suppo/delete",
+            'backURL'=>"/user2suppo/list/",
         ];
-        return view('user2suppo.add',$args);
 
+        return view('user2suppo.edit',$args)->with(parent::getAlertSessions());
     }
     //user2サポーター登録画面のPOST
-    public function u2screate(Request $request){
-        $this->validate($request, User2Supporter::$rules);
+    public function u2screate(User2SupporterRequest $request){
         $u2s = new User2Supporter;
         $form = $request->all();
         unset($form['_token']);
         $u2s->fill($form);
-
 
         $u2s->save();
 
@@ -173,19 +186,23 @@ class SupporterController extends Controller
         $args=[
         ];
 
-        return redirect()->route('u2s-add',$args)->with('alertComp',MessageConst::ADD_COMPLETED);
+        return redirect()->route('u2s-list',$args)->with('alertComp',MessageConst::ADD_COMPLETED);
 
     }
-    //
+    //user2サポーター削除のPOST
     public function u2sdelete(Request $request){
         $u2s_id = $request->u2s_id;
 
         $u2s = User2Supporter::find($u2s_id);
-        $u2s->delete();
+        if($u2s){
+            $u2s->delete();
+        }
 
         $args=[
         ];
-        return redirect()->route('u2s-add',$args)->with('alertComp',MessageConst::DELETE_COMPLETED);
+
+        return redirect()->route('u2s-list',$args)
+        ->with('alertComp',MessageConst::DELETE_COMPLETED);
     }
-    
+
 }
